+++ continuous_integration/scripts/create_json_schema.py
@@ -11,11 +11,12 @@ import functools
 import importlib
 import inspect
 import textwrap
-import typing as t
 from collections import defaultdict
+from collections.abc import Callable, Iterator, Mapping, Sequence
 from dataclasses import dataclass
 from graphlib import TopologicalSorter
 from pathlib import Path
+from typing import Any, Union, get_args, get_origin
 
 import click
 from boltons.strutils import under2camel
@@ -29,20 +30,20 @@ from pyxel import __version__
 @dataclass
 class Param:
     description: str
-    annotation: t.Any
+    annotation: Any
 
 
 @dataclass
 class ParamDefault:
     description: str
-    annotation: t.Any
-    default: t.Any
+    annotation: Any
+    default: Any
 
 
 @dataclass
 class FuncDocumentation:
     description: str
-    parameters: t.Mapping[str, t.Union[Param, ParamDefault]]
+    parameters: Mapping[str, Param | ParamDefault]
 
 
 @dataclass
@@ -50,7 +51,7 @@ class ModelInfo:
     model_name: str
     model_fullname: str
     model_class_name: str
-    func: t.Callable
+    func: Callable
 
 
 @dataclass
@@ -61,8 +62,8 @@ class ModelGroupInfo:
 
 @dataclass(frozen=True)
 class Klass:
-    cls: t.Type
-    base_cls: t.Optional[t.Type] = None
+    cls: type
+    base_cls: type | None = None
 
     @property
     def name(self) -> str:
@@ -70,7 +71,7 @@ class Klass:
 
 
 def get_annotation(annotation) -> str:
-    if t.get_origin(annotation):
+    if get_origin(annotation):
         annotation = str(annotation)
     elif hasattr(annotation, "__name__"):
         annotation = annotation.__name__
@@ -80,29 +81,50 @@ def get_annotation(annotation) -> str:
     return annotation
 
 
-def get_documentation(func: t.Callable) -> FuncDocumentation:
+def get_documentation(func: Callable) -> FuncDocumentation:
     assert func.__doc__
     doc = NumpyDocString(inspect.cleandoc(func.__doc__))
 
-    signature = inspect.signature(func)  # type: inspect.Signature
+    signature: inspect.Signature = inspect.signature(func)
 
     parameters = {}
 
     if doc["Parameters"]:
+        # Sanity checks
+        all_signature_params: set[str] = set(signature.parameters)
+        all_doc_params: set[str] = {doc_param.name for doc_param in doc["Parameters"]}
+
+        for doc_param in all_doc_params:
+            if ":" in doc_param:
+                raise RuntimeError(
+                    f"Bad format for {doc_param=}, for function {func=}."
+                )
+
+        if not all_doc_params.issubset(all_signature_params):
+            missing_params: set[str] = all_doc_params.difference(all_signature_params)
+            raise RuntimeError(
+                f"Missing key(s) in the signature: {', '.join(missing_params)} for function {func=}."
+            )
+
         for params in doc["Parameters"]:
             name, *_ = params.name.split(":")
             description = "\n".join(params.desc)
 
-            parameter = signature.parameters[name]  # type: inspect.Parameter
+            if name not in signature.parameters:
+                raise KeyError(
+                    f"Missing signature for parameter {name=} in function {func=}."
+                )
 
-            annotation = get_annotation(parameter.annotation)  # type: str
+            parameter: inspect.Parameter = signature.parameters[name]
+
+            annotation: str = get_annotation(parameter.annotation)
 
             if parameter.default != inspect.Parameter.empty:
-                param = ParamDefault(
+                param: ParamDefault | Param = ParamDefault(
                     description=description,
                     annotation=annotation.replace("NoneType", "None"),
                     default=parameter.default,
-                )  # type: t.Union[ParamDefault, Param]
+                )
             else:
                 param = Param(description=description, annotation=annotation)
 
@@ -110,11 +132,11 @@ def get_documentation(func: t.Callable) -> FuncDocumentation:
     else:
         for name, parameter in signature.parameters.items():
             if parameter.default != inspect.Parameter.empty:
-                param = ParamDefault(
+                param: ParamDefault | Param = ParamDefault(
                     description="",
                     annotation=parameter.annotation,
                     default=parameter.default,
-                )  # type: t.Union[ParamDefault, Param]
+                )
             else:
                 param = Param(description="", annotation=parameter.annotation)
 
@@ -128,10 +150,10 @@ def get_documentation(func: t.Callable) -> FuncDocumentation:
 @functools.cache
 def get_doc_from_klass(klass: Klass) -> FuncDocumentation:
     if klass.base_cls is None:
-        doc = get_documentation(klass.cls)  # type: FuncDocumentation
+        doc: FuncDocumentation = get_documentation(klass.cls)
     else:
-        doc_base = get_documentation(klass.base_cls)  # type: FuncDocumentation
-        doc_inherited = get_documentation(klass.cls)  # type: FuncDocumentation
+        doc_base: FuncDocumentation = get_documentation(klass.base_cls)
+        doc_inherited: FuncDocumentation = get_documentation(klass.cls)
 
         doc = FuncDocumentation(
             description=doc_inherited.description,
@@ -141,11 +163,11 @@ def get_doc_from_klass(klass: Klass) -> FuncDocumentation:
     return doc
 
 
-def generate_class(klass: Klass) -> t.Iterator[str]:
+def generate_class(klass: Klass) -> Iterator[str]:
     assert isinstance(klass, Klass)
 
     doc = get_doc_from_klass(klass)
-    klass_description_lst = textwrap.wrap(doc.description)  # type: t.Sequence[str]
+    klass_description_lst: Sequence[str] = textwrap.wrap(doc.description)
 
     yield "@schema("
     if len(klass_description_lst) == 1:
@@ -172,14 +194,14 @@ def generate_class(klass: Klass) -> t.Iterator[str]:
         for name, param in doc.parameters.items():
             title = name
 
-            if (origin := t.get_origin(param.annotation)) is not None:
-                args = t.get_args(param.annotation)  # type: t.Sequence
+            if (origin := get_origin(param.annotation)) is not None:
+                args: Sequence = get_args(param.annotation)
 
-                if origin == t.Union:
+                if origin == Union:
                     if len(args) != 2:
                         raise NotImplementedError
 
-                    annotation = f"Optional[{args[0].__name__}]"  # type: str
+                    annotation: str = f"Optional[{args[0].__name__}]"
                 else:
                     raise NotImplementedError
             else:
@@ -192,7 +214,7 @@ def generate_class(klass: Klass) -> t.Iterator[str]:
                 yield f"        default={param.default!r},"
             yield "        metadata=schema("
 
-            description_lst = textwrap.wrap(param.description)  # type: t.Sequence[str]
+            description_lst: Sequence[str] = textwrap.wrap(param.description)
             if len(description_lst) == 1:
                 yield f"            title={title!r},"
                 yield f"            description={description_lst[0]!r}"
@@ -216,13 +238,13 @@ def generate_class(klass: Klass) -> t.Iterator[str]:
 
 
 def generate_model(
-    func: t.Callable,
+    func: Callable,
     func_name: str,
     func_fullname: str,
     model_name: str,
-) -> t.Iterator[str]:
+) -> Iterator[str]:
 
-    doc = get_documentation(func)  # type: FuncDocumentation
+    doc: FuncDocumentation = get_documentation(func)
 
     yield "@schema(title='Parameters')"
     yield "@dataclass"
@@ -230,9 +252,7 @@ def generate_model(
 
     dct = {key: value for key, value in doc.parameters.items() if key != "detector"}
 
-    all_defaults = all(
-        [isinstance(el, ParamDefault) for el in dct.values()]
-    )  # type: bool
+    all_defaults: bool = all([isinstance(el, ParamDefault) for el in dct.values()])
 
     if dct:
         for name, param in dct.items():
@@ -246,7 +266,7 @@ def generate_model(
             yield "        metadata=schema("
             yield f"            title={title!r}"
 
-            description_lst = textwrap.wrap(param.description)  # type: t.Sequence[str]
+            description_lst: Sequence[str] = textwrap.wrap(param.description)
             if len(description_lst) == 1:
                 yield f"            ,description={description_lst[0]!r}"
             elif len(description_lst) > 1:
@@ -301,19 +321,19 @@ def generate_model(
     yield ""
 
 
-def get_model_info(group_name: str) -> t.Sequence[ModelInfo]:
+def get_model_info(group_name: str) -> Sequence[ModelInfo]:
     group_module = importlib.import_module(f"pyxel.models.{group_name}")
 
-    lst = []  # type: t.List[ModelInfo]
+    lst: list[ModelInfo] = []
     for name in dir(group_module):
         if name.startswith("__"):
             continue
 
-        func = getattr(group_module, name)  # type: t.Callable
+        func: Callable = getattr(group_module, name)
         if not callable(func):
             continue
 
-        sig = inspect.signature(func)  # type: inspect.Signature
+        sig: inspect.Signature = inspect.signature(func)
 
         if "detector" not in sig.parameters:
             continue
@@ -334,14 +354,15 @@ def capitalize_title(name: str) -> str:
     return " ".join([el.capitalize() for el in name.split("_")])
 
 
-def generate_group(model_groups_info: t.Sequence[ModelGroupInfo]) -> t.Iterator[str]:
+def generate_group(model_groups_info: Sequence[ModelGroupInfo]) -> Iterator[str]:
 
     for group_info in model_groups_info:
         group_name = group_info.name
 
-        models_info = get_model_info(group_name)  # type: t.Sequence[ModelInfo]
+        models_info: Sequence[ModelInfo] = get_model_info(group_name)
 
-        for info in models_info:  # type: ModelInfo
+        info: ModelInfo
+        for info in models_info:
 
             group_name_title = capitalize_title(group_name)
             model_title = capitalize_title(info.model_name)
@@ -377,7 +398,7 @@ def generate_group(model_groups_info: t.Sequence[ModelGroupInfo]) -> t.Iterator[
         yield f"    ] = field(default=None, metadata=schema(title={group_name_title!r}))"
 
 
-def get_model_group_info() -> t.Sequence[ModelGroupInfo]:
+def get_model_group_info() -> Sequence[ModelGroupInfo]:
     all_group_models = (
         "photon_generation",
         "optics",
@@ -390,44 +411,42 @@ def get_model_group_info() -> t.Sequence[ModelGroupInfo]:
         "signal_transfer",
     )
 
-    lst = []  # type: t.List[ModelGroupInfo]
-
-    for group_name in all_group_models:
-        lst.append(
-            ModelGroupInfo(
-                name=group_name, class_name=under2camel(f"model_group_{group_name}")
-            )
+    lst = [
+        ModelGroupInfo(
+            name=group_name,
+            class_name=under2camel(f"model_group_{group_name}"),
         )
+        for group_name in all_group_models
+    ]
 
     return lst
 
 
 @functools.cache
-def create_klass(cls: t.Union[t.Type, str]) -> Klass:
+def create_klass(cls: type | str) -> Klass:
     import pyxel.detectors
 
     if isinstance(cls, str):
-        cls_type = getattr(pyxel.detectors, cls)  # type: t.Type
+        cls_type: type = getattr(pyxel.detectors, cls)
         return create_klass(cls_type)
 
     # Try to find a base class
-    if (origin := t.get_origin(cls)) is not None:
-        args = t.get_args(cls)  # type: t.Sequence
+    if (origin := get_origin(cls)) is not None:
+        args: Sequence = get_args(cls)
 
-        match origin:
-            case t.Union:
-                if len(args) != 2:
-                    raise NotImplementedError
-
-                # Optional type
-                klass = args[0]
-
-            case _:
+        if origin == Union:
+            if len(args) != 2:
                 raise NotImplementedError
+
+            # Optional type
+            klass = args[0]
+
+        else:
+            raise NotImplementedError
     else:
         klass = cls
 
-    _, *base_classes, _ = inspect.getmro(klass)  # type: tuple
+    _, *base_classes, _ = inspect.getmro(klass)
 
     if base_classes:
         return Klass(klass, base_cls=base_classes[0])
@@ -435,43 +454,43 @@ def create_klass(cls: t.Union[t.Type, str]) -> Klass:
     return Klass(klass)
 
 
-def create_graph(cls: t.Type, graph: t.Mapping[Klass, t.Set[Klass]]) -> None:
-    klass = create_klass(cls)  # type: Klass
+def create_graph(cls: type, graph: Mapping[Klass, set[Klass]]) -> None:
+    klass: Klass = create_klass(cls)
 
-    doc = get_doc_from_klass(klass)  # type: FuncDocumentation
+    doc: FuncDocumentation = get_doc_from_klass(klass)
 
     if klass.base_cls is None:
-        parameters = (
-            doc.parameters
-        )  # type: t.Mapping[str, t.Union[Param, ParamDefault]]
+        parameters: Mapping[str, Param | ParamDefault] = doc.parameters
     else:
         create_graph(cls=klass.base_cls, graph=graph)
-        klass_base = create_klass(klass.base_cls)  # type: Klass
+        klass_base: Klass = create_klass(klass.base_cls)
         graph[klass].add(klass_base)
 
-        klass_base_doc = get_doc_from_klass(klass_base)  # type:  FuncDocumentation
+        klass_base_doc: FuncDocumentation = get_doc_from_klass(klass_base)
 
         parameters = {**doc.parameters, **klass_base_doc.parameters}
 
     for name, parameter in parameters.items():
         assert parameter.annotation
 
-        klass_param = create_klass(parameter.annotation)  # type: Klass
+        klass_param: Klass = create_klass(parameter.annotation)
         graph[klass].add(klass_param)
 
         if klass_param.base_cls is not None:
-            klass_param_base = create_klass(klass_param.base_cls)  # type: Klass
+            klass_param_base: Klass = create_klass(klass_param.base_cls)
             graph[klass_param].add(klass_param_base)
 
 
-def generate_detectors() -> t.Iterator[str]:
+def generate_detectors() -> Iterator[str]:
     from pyxel.detectors import APD, CCD, CMOS, MKID, Detector
 
-    registered_detectors = (CCD, CMOS, MKID, APD)  # type: t.Sequence[t.Type[Detector]]
+    registered_detectors: Sequence[type[Detector]] = (CCD, CMOS, MKID, APD)
 
     # Build a dependency graph
-    graph = defaultdict(set)  # type: t.Mapping[Klass, t.Set[Klass]]
-    for detector in registered_detectors:  # type: t.Type[Detector]
+    graph: Mapping[Klass, set[Klass]] = defaultdict(set)
+
+    detector: type[Detector]
+    for detector in registered_detectors:
         create_graph(cls=detector, graph=graph)
 
     # Generate code based on the dependency graph
@@ -655,7 +674,7 @@ def generate_detectors() -> t.Iterator[str]:
     yield ""
 
     # Create wrappers for the detectors
-    detector_classes = ["CCD", "CMOS", "MKID", "APD"]  # type: t.Sequence[str]
+    detector_classes: Sequence[str] = ("CCD", "CMOS", "MKID", "APD")
     # for klass_name in detector_classes:
     #     yield f"#@schema(title='{klass_name}')"
     #     yield "#@dataclass"
@@ -665,7 +684,7 @@ def generate_detectors() -> t.Iterator[str]:
     #     yield ""
 
     # Create wrappers for the modes
-    mode_classes = ["Exposure", "Observation", "Calibration"]  # type: t.Sequence[str]
+    mode_classes: Sequence[str] = ("Exposure", "Observation", "Calibration")
     # for klass_name in mode_classes:
     #     yield f"#@schema(title='{klass_name}')"
     #     yield "#@dataclass"
@@ -694,7 +713,7 @@ def generate_detectors() -> t.Iterator[str]:
         yield f"    {klass_name.lower()}_detector: Optional[{klass_name}] = field(default=None, metadata=schema(title={klass_name!r}))"
 
 
-def generate_all_models() -> t.Iterator[str]:
+def generate_all_models() -> Iterator[str]:
     lst = get_model_group_info()
     yield "#  Copyright (c) European Space Agency, 2017, 2018, 2019, 2020, 2021, 2022."
     yield "#"
@@ -709,15 +728,18 @@ def generate_all_models() -> t.Iterator[str]:
     yield "######################################"
     yield ""
 
+    yield "import json"
     yield "import pathlib"
     yield "import sys"
-    yield "from typing import Optional, Mapping, Any, Sequence, Iterator, Literal, Union, Tuple"
     yield "from dataclasses import dataclass, field"
+    yield "from pathlib import Path"
+    yield "from pprint import pprint"
+    yield "from typing import Any, Iterator, Literal, Mapping, Optional, Sequence, Tuple, Union"
+    yield ""
     yield "import click"
     yield "from apischema import schema"
-    yield "import json"
-    yield "from pathlib import Path"
     yield "from apischema.json_schema import JsonSchemaVersion, deserialization_schema"
+    yield "from deepdiff import DeepDiff"
 
     yield ""
     yield ""
@@ -764,7 +786,11 @@ def generate_all_models() -> t.Iterator[str]:
     yield "    help='JSON schema filename',"
     yield "    show_default=True,"
     yield ")"
-    yield "def create_json_schema(filename: pathlib.Path):"
+
+    yield "@click.option('--check', is_flag=True,"
+    yield '     help="Don\'t write the JSON Schema back, just return the status.")'
+
+    yield "def create_json_schema(filename: pathlib.Path, check: bool):"
     yield "    if sys.version_info < (3, 10):"
     yield "        raise NotImplementedError('This script must run on Python 3.10+')"
     yield ""
@@ -775,12 +801,30 @@ def generate_all_models() -> t.Iterator[str]:
     yield "        Configuration, version=JsonSchemaVersion.DRAFT_7, all_refs=True,"
     yield "    )"
     yield ""
-    yield "    print(json.dumps(dct_schema))"
     yield ""
     yield "    full_filename = pathlib.Path(filename).resolve()"
     yield ""
-    yield "    with full_filename.open('w') as fh:"
-    yield "        json.dump(obj=dct_schema, fp=fh, indent=2)"
+    yield "    if check:"
+    yield "        with full_filename.open() as fh:"
+    yield "            dct_reference = json.load(fh)"
+    yield ""
+    yield "        new_dct_schema: Mapping[str, Any] = json.loads(json.dumps(dct_schema))"
+    yield ""
+    yield "        if dct_reference == new_dct_schema:"
+    yield "            sys.exit(0)"
+    yield "        else:"
+    yield "            result = DeepDiff(dct_reference, new_dct_schema)"
+    yield ""
+    yield "            print("
+    yield '                f"Error, JSON Schema file: {full_filename} is not the newest version. "'
+    yield "                f\"Please run 'tox -e json_schema'\""
+    yield "             )"
+    yield "            pprint(result)"
+    yield "            sys.exit(1)"
+    yield "    else:"
+    yield "        print(json.dumps(dct_schema))"
+    yield "        with full_filename.open('w') as fh:"
+    yield "            json.dump(obj=dct_schema, fp=fh, indent=2)"
     yield ""
     yield ""
     yield "if __name__ == '__main__':"